Add one-click-embedding wrapper for gist
